drv-i2c-devices: make TMP117 driver nicer
This code appears to predate `u16::from_be_bytes`, and therefore is in
the habit of representing a big-endian 16 bit register as a tuple and
then passing it around, which I found unfortunate. Presumably the fact
that only the driver understands the tuple's endianness is also the
reason it, it makes the enum of device registers public, but does not
make the `read_reg` function public, which I *also* found unfortunate.
THis commit fixes all of that. It also changes `host=-sp-comms` to
actually use the driver instead of ignoring it.
