+++ drv/i2c-devices/src/tmp117.rs
@@ -40,8 +40,8 @@ pub struct Tmp117 {
     device: I2cDevice,
 }
 
-fn convert(raw: (u8, u8)) -> Celsius {
-    Celsius(f32::from(i16::from(raw.0) << 8 | i16::from(raw.1)) / 128.0)
+fn convert(raw: u16) -> Celsius {
+    Celsius(f32::from(raw as i16) / 128.0)
 }
 
 impl core::fmt::Display for Tmp117 {
@@ -55,19 +55,11 @@ impl Tmp117 {
         Self { device: *device }
     }
 
-    fn read_reg(&self, reg: Register) -> Result<(u8, u8), Error> {
+    pub fn read_reg(&self, reg: Register) -> Result<u16, Error> {
         match self.device.read_reg::<u8, [u8; 2]>(reg as u8) {
-            Ok(buf) => Ok((buf[0], buf[1])),
+            Ok(buf) => Ok(u16::from_be_bytes(buf)),
             Err(code) => Err(Error::BadRegisterRead { reg, code }),
         }
-    }
-
-    pub fn read_eeprom(&self) -> Result<[u8; 6], Error> {
-        let ee1 = self.read_reg(Register::EEPROM1)?;
-        let ee2 = self.read_reg(Register::EEPROM2)?;
-        let ee3 = self.read_reg(Register::EEPROM3)?;
-
-        Ok([ee1.0, ee1.1, ee2.0, ee2.1, ee3.0, ee3.1])
     }
 }
 
@@ -75,7 +67,7 @@ impl Validate<Error> for Tmp117 {
     fn validate(device: &I2cDevice) -> Result<bool, Error> {
         let id = Tmp117::new(device).read_reg(Register::DeviceID)?;
 
-        Ok(id.0 == 0x1 && id.1 == 0x17)
+        Ok(id == 0x0117)
     }
 }
